Removing editable attributes from ObsReferenceRange
