+++ api/src/main/java/org/openmrs/BaseReferenceRange.java
@@ -9,9 +9,6 @@
  */
 package org.openmrs;
 
-import org.codehaus.jackson.annotate.JsonIgnore;
-import org.hibernate.search.annotations.Field;
-
 import javax.persistence.Column;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -23,7 +20,7 @@ import java.util.Date;
  * It contains all reference ranges.
  */
 @MappedSuperclass
-public abstract class BaseReferenceRangeObject extends BaseOpenmrsObject implements Auditable, Retireable {
+public abstract class BaseReferenceRange extends BaseOpenmrsObject {
 	@Column(name = "hi_absolute")
 	private Double hiAbsolute;
 
@@ -46,28 +43,8 @@ public abstract class BaseReferenceRangeObject extends BaseOpenmrsObject impleme
 	@JoinColumn(name = "creator")
 	private User creator;
 
-	@ManyToOne
-	@JoinColumn(name = "changed_by")
-	private User changedBy;
-
-	@Column(name = "date_changed")
-	private Date dateChanged;
-
 	@Column(name = "date_created")
 	private Date dateCreated;
-	
-	@ManyToOne
-	@JoinColumn(name = "retired_by")
-	private User retiredBy;
-
-	@Column(name = "date_retired")
-	private Date dateRetired;
-
-	@Column(name = "retire_reason")
-	private String retireReason;
-
-	@Field
-	private Boolean retired = false;
 
 	// Property accessors
 
@@ -180,41 +157,8 @@ public abstract class BaseReferenceRangeObject extends BaseOpenmrsObject impleme
 	}
 
 	/**
-	 * @return Returns the changedBy.
-	 */
-	@Override
-	public User getChangedBy() {
-		return changedBy;
-	}
-
-	/**
-	 * @param changedBy The changedBy to set.
-	 */
-	@Override
-	public void setChangedBy(User changedBy) {
-		this.changedBy = changedBy;
-	}
-
-	/**
-	 * @return Returns the dateChanged.
-	 */
-	@Override
-	public Date getDateChanged() {
-		return dateChanged;
-	}
-
-	/**
-	 * @param dateChanged The dateChanged to set.
-	 */
-	@Override
-	public void setDateChanged(Date dateChanged) {
-		this.dateChanged = dateChanged;
-	}
-
-	/**
 	 * @return Returns the creator.
 	 */
-	@Override
 	public User getCreator() {
 		return creator;
 	}
@@ -222,7 +166,6 @@ public abstract class BaseReferenceRangeObject extends BaseOpenmrsObject impleme
 	/**
 	 * @param creator The creator to set.
 	 */
-	@Override
 	public void setCreator(User creator) {
 		this.creator = creator;
 	}
@@ -231,7 +174,6 @@ public abstract class BaseReferenceRangeObject extends BaseOpenmrsObject impleme
 	/**
 	 * @return Returns the dateCreated.
 	 */
-	@Override
 	public Date getDateCreated() {
 		return dateCreated;
 	}
@@ -239,88 +181,7 @@ public abstract class BaseReferenceRangeObject extends BaseOpenmrsObject impleme
 	/**
 	 * @param dateCreated The dateCreated to set.
 	 */
-	@Override
 	public void setDateCreated(Date dateCreated) {
 		this.dateCreated = dateCreated;
 	}
-	
-	/**
-	 * @return Returns the retired.
-	 *
-	 * @deprecated as of 2.0, use {@link #getRetired()}
-	 */
-	@Override
-	@Deprecated
-	@JsonIgnore
-	public Boolean isRetired() {
-		return getRetired();
-	}
-
-	/**
-	 * This method delegates to {@link #isRetired()}.
-	 *
-	 * @see org.openmrs.Retireable#isRetired()
-	 */
-	@Override
-	public Boolean getRetired() {
-		return retired;
-	}
-
-	/**
-	 * @param retired The retired to set.
-	 */
-	@Override
-	public void setRetired(Boolean retired) {
-		this.retired = retired;
-	}
-
-	/**
-	 * @return the retiredBy
-	 */
-	@Override
-	public User getRetiredBy() {
-		return retiredBy;
-	}
-
-	/**
-	 * @param retiredBy the retiredBy to set
-	 */
-	@Override
-	public void setRetiredBy(User retiredBy) {
-		this.retiredBy = retiredBy;
-	}
-
-	/**
-	 * @return the dateRetired
-	 */
-	@Override
-	public Date getDateRetired() {
-		return dateRetired;
-	}
-
-	/**
-	 * @param dateRetired the dateRetired to set
-	 */
-	@Override
-	public void setDateRetired(Date dateRetired) {
-		this.dateRetired = dateRetired;
-	}
-
-	/**
-	 * @return the retireReason
-	 */
-	@Override
-	public String getRetireReason() {
-		return retireReason;
-	}
-
-	/**
-	 * @param retireReason the retireReason to set
-	 */
-	@Override
-	public void setRetireReason(String retireReason) {
-		this.retireReason = retireReason;
-	}
-
-
 }
