+++ api/src/main/java/org/openmrs/ConceptReferenceRange.java
@@ -9,8 +9,10 @@
  */
 package org.openmrs;
 
+import org.codehaus.jackson.annotate.JsonIgnore;
 import org.hibernate.envers.Audited;
 import org.hibernate.search.annotations.DocumentId;
+import org.hibernate.search.annotations.Field;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -21,6 +23,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+import java.util.Date;
 
 /**
  * A concept reference range is typically a range of a concept for certain factor(s) e.g. age, gender e.t.c. 
@@ -35,7 +38,7 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "concept_reference_range")
 @Audited
-public class ConceptReferenceRange extends BaseReferenceRangeObject {
+public class ConceptReferenceRange extends BaseReferenceRange implements Auditable, Retireable {
 	private static final long serialVersionUID = 47329L;
 
 	// Fields
@@ -51,6 +54,30 @@ public class ConceptReferenceRange extends BaseReferenceRangeObject {
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "concept_id", nullable = false)
 	private Concept concept;
+
+	@ManyToOne
+	@JoinColumn(name = "retired_by")
+	private User retiredBy;
+
+	@Column(name = "date_retired")
+	private Date dateRetired;
+
+	@Column(name = "retire_reason")
+	private String retireReason;
+
+	@Field
+	private Boolean retired = false;
+
+	@ManyToOne
+	@JoinColumn(name = "changed_by")
+	private User changedBy;
+
+	@Column(name = "date_changed")
+	private Date dateChanged;
+
+	@ManyToOne
+	@JoinColumn(name = "creator")
+	private User creator;
 	
 	// Constructors
 	
@@ -122,5 +149,129 @@ public class ConceptReferenceRange extends BaseReferenceRangeObject {
 	@Override
 	public void setId(Integer id) {
 		setConceptReferenceRangeId(id);
+	}
+
+	/**
+	 * @return Returns the retired.
+	 *
+	 * @deprecated as of 2.0, use {@link #getRetired()}
+	 */
+	@Override
+	@Deprecated
+	@JsonIgnore
+	public Boolean isRetired() {
+		return getRetired();
+	}
+
+	/**
+	 * This method delegates to {@link #isRetired()}.
+	 *
+	 * @see org.openmrs.Retireable#isRetired()
+	 */
+	@Override
+	public Boolean getRetired() {
+		return retired;
+	}
+
+	/**
+	 * @param retired The retired to set.
+	 */
+	@Override
+	public void setRetired(Boolean retired) {
+		this.retired = retired;
+	}
+
+	/**
+	 * @return the retiredBy
+	 */
+	@Override
+	public User getRetiredBy() {
+		return retiredBy;
+	}
+
+	/**
+	 * @param retiredBy the retiredBy to set
+	 */
+	@Override
+	public void setRetiredBy(User retiredBy) {
+		this.retiredBy = retiredBy;
+	}
+
+	/**
+	 * @return the dateRetired
+	 */
+	@Override
+	public Date getDateRetired() {
+		return dateRetired;
+	}
+
+	/**
+	 * @param dateRetired the dateRetired to set
+	 */
+	@Override
+	public void setDateRetired(Date dateRetired) {
+		this.dateRetired = dateRetired;
+	}
+
+	/**
+	 * @return the retireReason
+	 */
+	@Override
+	public String getRetireReason() {
+		return retireReason;
+	}
+
+	/**
+	 * @param retireReason the retireReason to set
+	 */
+	@Override
+	public void setRetireReason(String retireReason) {
+		this.retireReason = retireReason;
+	}
+
+	/**
+	 * @return Returns the changedBy.
+	 */
+	@Override
+	public User getChangedBy() {
+		return changedBy;
+	}
+
+	/**
+	 * @param changedBy The changedBy to set.
+	 */
+	@Override
+	public void setChangedBy(User changedBy) {
+		this.changedBy = changedBy;
+	}
+
+	/**
+	 * @return Returns the dateChanged.
+	 */
+	@Override
+	public Date getDateChanged() {
+		return dateChanged;
+	}
+
+	/**
+	 * @param dateChanged The dateChanged to set.
+	 */
+	@Override
+	public void setDateChanged(Date dateChanged) {
+		this.dateChanged = dateChanged;
+	}
+
+	/**
+	 * @return Returns the creator.
+	 */
+	public User getCreator() {
+		return creator;
+	}
+
+	/**
+	 * @param creator The creator to set.
+	 */
+	public void setCreator(User creator) {
+		this.creator = creator;
 	}
 }
