+++ api/src/test/resources/org/openmrs/include/standardTestDataset.xml
@@ -449,5 +449,5 @@
 	<concept_reference_range concept_reference_range_id="30" concept_id="25" criteria="#set( $criteria = $patient.getAge() &gt;= 1 &amp;&amp; $patient.getAge() &lt;= 3 )$criteria" hi_absolute="150.0" hi_normal="118.0" low_normal="80" low_absolute="80" creator="1" date_created="2024-01-01 00:00:00.0" retired="0" uuid="239c1904-15ff-45e1-ac9d-d83afb637926"/>
 	<concept_reference_range concept_reference_range_id="31" concept_id="26" criteria="#set( $criteria = $patient.getAge() &gt;= 1 &amp;&amp; $patient.getAge() &lt;= 3 )$criteria" hi_absolute="120.0" hi_normal="118.0" low_normal="80" low_absolute="60" creator="1" date_created="2024-01-01 00:00:00.0" retired="0" uuid="5269e035-c8e6-44d7-af47-7a25a6cf6c60"/>
 	<concept_reference_range concept_reference_range_id="32" concept_id="27" criteria="#set( $criteria = $patient.getAge() &gt;= 1 &amp;&amp; $patient.getAge() &lt;= 10 )$criteria" hi_absolute="120.0" hi_normal="118.0" low_normal="80" low_absolute="60" creator="1" date_created="2024-01-01 00:00:00.0" retired="0" uuid="d8ad3ceb-00a4-4484-a1ff-700f7ac3d10d"/>
-	<obs_reference_range obs_reference_range_id="1000" obs_id="7" hi_absolute="120.0" hi_normal="118.0" low_normal="80" low_absolute="60" creator="1" date_created="2024-01-01 00:00:00.0" retired="0" uuid="099b6366-4b2a-421c-8c0b-959e2bc75450"/>
+<!--	<obs_reference_range obs_reference_range_id="1000" obs_id="70000" hi_absolute="120.0" hi_normal="118.0" low_normal="80" low_absolute="60" creator="1" date_created="2024-01-01 00:00:00.0" uuid="099b6366-4b2a-421c-8c0b-959e2bc75450"/>-->
 </dataset>
